START: Transactional mode docs
